Separate files for each class and globbing on CMakeLists.txt
