+++ src/includes/ElectricLib.h
@@ -7,7 +7,7 @@
 #define ELECTRICLIB_H
 
 #include <H5Cpp.h>
-#include "TethysLib.h"
+#include "TethysBaseLib.h"
 #include "TethysMathLib.h"
 #include "Tethys1DLib.h"
 #include "Tethys2DLib.h"
