+++ src/includes/Tethys2DLib.h
@@ -7,8 +7,9 @@
 #define TETHYS2DLIB_H
 
 #include <H5Cpp.h>
-#include "TethysLib.h"
+#include "TethysBaseLib.h"
 #include "TethysMathLib.h"
+#include "SetUpParametersLib.h"
 
 using namespace H5;
