+++ src/includes/TethysBaseLib.h
@@ -3,8 +3,8 @@
 * Distributed under the MIT License (license terms are at http://opensource.org/licenses/MIT).   *
 \************************************************************************************************/
 
-#ifndef TETHYSLIB_H
-#define TETHYSLIB_H
+#ifndef TETHYSBASELIB_H
+#define TETHYSBASELIB_H
 
 #include <fstream>
 #include <cstdio>
@@ -35,34 +35,6 @@ using namespace H5;
 
 const FloatType      HDF5FLOAT(PredType::NATIVE_FLOAT);
 const IntType        HDF5INT(PredType::NATIVE_INT);
-
-
-/*!
- * @brief Initialization class for the fluid classes.
- *
- * This class allows the initialization of the appropriate physical parameters either from inline arguments, prompt or reading them from an existing HDF5 file
- * */
-class SetUpParameters {
-	public:
-		SetUpParameters();
-		SetUpParameters(int argc, char ** argv);
-		SetUpParameters(float sound, float fermi, float coll, float visco, float cyclo, int mode, float aspect);
-		~SetUpParameters() = default;
-		int SaveMode;
-		int SizeX;
-		int SizeY;
-		float Length=1.0f;
-		float Width=1.0f;
-		float AspectRatio=1.0f;
-		float SoundVelocity;
-		float FermiVelocity;
-		float CollisionFrequency;
-		float ShearViscosity;
-		float CyclotronFrequency;
-		void ExceptionsChecking() const; ///< Runs a checking on the physical feasibility of the parameters
-		void DefineGeometry(); ///< Set ups the 2D grid dimensions
-		void ParametersFromHdf5File(const std::string& hdf5name); ///< Imports the parameters from a saved HDF5 file
-};
 
 
 /*!
